% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\docType{data}
\name{A}
\alias{A}
\title{A matrix}
\format{A matrix of size n x n, where n = 87
(morphological traits, the 87 coordinates):
\describe{
  \item{n}{Number of coordinates for which the correlation values
  have been calculated, 87}
}}
\usage{
A
}
\description{
A matrix which, once multiplied by its transpose,
}
\keyword{datasets}
