% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{array2matrix}
\alias{array2matrix}
\title{Convert an array into a matrix}
\usage{
array2matrix(X, coords = c(2, 3))
}
\arguments{
\item{X}{Array, \code{k} landmark points, \code{q} coordinates, and
\code{s} specimens.}

\item{coords}{Integer, 2 or 3, for 2D or 3D landmarks, respectively.}
}
\value{
An object of class matrix, with \code{s} rows, one for specimen, and \code{n} columns, one
for each coordinate of the landmarks.
Each landmark can be given in 2D or 3D. For instance,
if the landmarks are 3D, the first 3 columns will be the
coordinates x, y, and z for the first landmark, the next 3
columns for the second landmark, and so on.
\tabular{cccccccc}{
 specimens \tab lmk1.x \tab lmk1.y  \tab lmk1.z  \tab lmk2.x \tab lmk2.y \tab lmk2.z  \tab ... \cr
 Sp_1      \tab 0.143  \tab -0.028  \tab -0.044  \tab 0.129  \tab 0.028  \tab -0.043  \tab ... \cr
 Sp_2      \tab 0.128  \tab -0.024  \tab -0.028  \tab 0.124  \tab 0.027  \tab -0.025  \tab ... \cr
 ...       \tab ...    \tab ...     \tab ...     \tab ...    \tab ...    \tab ...     \tab ...
}

See object \code{C} for a more detailed example of the format of the
object that is returned and \code{data-raw/C.R} for the explanation
about how to generate this object.

Note that if the names for the \code{s} matrices, one per specimen, in the array
are not provided, i.e. the names for specimens are not given,
the specimens in the returned matrix, \code{row.names(matrix)}, will be
labelled as '1', '2', and so on.
}
\description{
Convert an array with landmark points into an object of class matrix.
}
\details{
The object \code{X}, class array, has format \code{k x q x s}, where \code{k} is
the number of landmarks, \code{q} the number of coordinates, and \code{s} the number
of specimens. See \code{C.arr.unal} for an example of the format of a 3D array
and \code{data-raw/C.R} for the details about how to generate this object.
}
\seealso{
\code{\link{matrix2array}}, \code{\link{write.morpho}}
}
\author{
Sandra Alvarez-Carretero and Mario dos Reis
}
