% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-lhd.R
\name{block.boot}
\alias{block.boot}
\title{Generate block bootstrap replicates of sampled power likelihoods}
\usage{
block.boot(R, p = 0.1, mcmcf = "mcmc.txt", betaf = "beta.txt",
  preff = "lnL")
}
\arguments{
\item{R}{numeric, number of bootstrap replicates}

\item{p}{numeric, block length, giving as a proportion of the MCMC sample size}

\item{mcmcf}{character, mcmc output file name}

\item{betaf}{character, file with beta values}

\item{preff}{character, prefix for files storing boot replicates}
}
\description{
Generate block bootstrap replicates of sampled power likelihoods
}
\details{
Block bootstrap replicates are generated using the stationary boostrap method
of Politis and Romano (1994). The replicates are stored in files named using
\code{preff} and the replicate number. For example, if \code{preff = "lnL"}
(the default) then the files are lnL0.txt, lnL1.txt, lnL2.txt, ..., etc, with
lnL0.txt corresponding to the original log-likelihood sample. Replicates are
stored within the directories corresponding to the appropriate beta values.
The collection of files can grow large quickly so you may want to use a small
number of replicates (say R = 10 to R = 100).

This function uses code extracted from the boot package by Canty and Ripley.
}
\references{
Politis and Romano (1994) The stationary boostrap. \emph{J. Am. Stat. Assoc.}, 
89: 1303--1313.
}
\seealso{
\link{stepping.stones.boot} and \link{tsboot} (from the boot package).
}
