% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C.mat.unal.R
\docType{data}
\name{C.mat.unal}
\alias{C.mat.unal}
\title{29 3D landmarks from the skulls of 19 carnivoran specimens before Procrustes analysis}
\format{
A matrix with s = 19 rows and n = 87 columns (87/3 = 29 landmarks):
  \describe{
  \item{s}{Rows, specimens from which landmarks were collected, 19}
  \item{n}{Columns, 87 traits (29 landmarks in 3D) after the PA}
}
}
\usage{
C.mat.unal
}
\description{
A matrix containing the 29 3D landmarks collected from the skulls of 19 carnivoran
specimens before carrying out a Procrustes analysis (PA). Please take a look at the
description in morpho/data-raw/C.R to understand
how this object was generated.
}
\keyword{datasets}
