% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C.PS.R
\docType{data}
\name{C.PS}
\alias{C.PS}
\title{Object of class procSym output by Morpho after PA}
\format{
Object procSym
  \describe{
  \item{...}{Check \code{\link[Morpho]{procSym}} for more details}
}
}
\usage{
C.PS
}
\description{
Object of class procSym output by Morpho, which mean shape is later
used to align the foxes ("Vulpes vulpes") specimens to it. Please take a
look at the description in morpho/data-raw/C.R to understand
how this object was generated.
}
\keyword{datasets}
