% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{calc.pop.cor}
\alias{calc.pop.cor}
\title{Estimate shrunk correlation matrix from population matrix}
\usage{
calc.pop.cor(P, delta = 0.01)
}
\arguments{
\item{P}{Matrix, population matrix of size \code{s x n}, 's' specimens
and 'n' characters.}

\item{delta}{Numeric, shrinkage value used to generate the estimated
shrunk correlation matrix (see details).}
}
\value{
\item{P}{Matrix with the simulated population sample of
\code{psample} specimens and \code{n} morphological continuous traits
sampled from a normal distribution with mean = 0 and sd = sqrt(c)}
}
\description{
Estimate the shrunk correlation matrix from a population that could
have been either previously simulated with \code{\link{write.morpho}}
or could refer to a population sample from a real data set.
}
\details{
Thre is a problem commonly faced with morphological data because, usually,
the matrices with these data have more columns (more traits) than rows
(specimens from which the traits were sampled). As the unbiased correlation
matrices that can be calculated from these morphological matrices tend to
be singular, they are not invertible. Consequently, they cannot be used
to transform the data sets when accounting for trait correlation nor used to
calculate the corresponding determinants, needed during the likelihood
calculation in MCMCTree.

The \code{delta} parameter is a shrinkage value used to find the estimate
of the shrunk correlation matrix, which is to be symmetric and invertible,
thus it can be used to transform the morphological continuous data set
and during the likelihood calculation in MCMCTree. This estimate can be found
by using the following equation,
\eqn{{R}^{*} = \delta I + (1-\delta){R'}}{R.sh = delta I + (1-delta)R.unb},
as detailed in Schaffer & Strimmer, 2005.
}
\examples{

# Simulation setup: Estimate the shrunk correlation matrix
# of a sample population (20 specimens). The shrinkage value
# is the one also used as a default.

  morpho::calc.pop.cor( P = sim.population, delta = 0.01 )

}
\references{
\insertRef{Schafer2005}{morpho}
}
\seealso{
\code{\link{sim.morpho}}, \code{\link{write.morpho}}
}
\author{
Sandra Alvarez-Carretero and Mario dos Reis
}
