% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnivores.R
\docType{data}
\name{carnivores}
\alias{carnivores}
\title{Carnivores dataset}
\format{
\code{carnivores} is a list with elements:

\code{C.raw}, a matrix of 29 3D-landmarks measurements from 19 species;

\code{V.raw}, a matrix of 29 3D-landmarks for 21 common foxes (Vulpes
vulpes);

\code{C.proc} and \code{V.proc}, the corresponding matrices after Procrustes
alignment;

\code{var.foxes} and \code{R.sh}, with estimates of the variance vector and
correlation matrix for the landmarks in the 21 foxes;

\code{M}, a matrix of transformed landmarks (using the foxes variances
and correlation matrix, with Cholesky metric) for the 19 carnivores;

\code{pairedLM}, a matrix describing the symmetry of the landmarks;

\code{tree}, the phylogeny of the 19 carnivores; and

\code{mcmc}, an MCMC sample of divergence times and morphological and
molecular rates from an MCMCtree analysis.
}
\source{
Alvarez-Carretero S, Goswami A, Yang Z and dos Reis M. (2018) Bayesian
estimation of species divergence times using correlated quantitative
characters. \emph{Syst. Biol.,} 68: 967--986.
}
\usage{
carnivores
}
\description{
Dataset of carnivores containing alignments of morphlogical continuous
characters, a phylogeny, and an MCMC matrix produced by MCMCtree.
}
\keyword{datasets}
