% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.raw.R
\docType{data}
\name{coords.raw}
\alias{coords.raw}
\title{48 3D landmarks from the skulls of 8 Triturius specimens and 1 Calotriton specimen}
\format{A data.frame with n = 9 rows and p = 145 columns (info column + 144 landamrks):
\describe{
  \item{n}{Rows, speciments from which landmarks were collected, 9}
  \item{p}{Columns, information about the specimens set (1st column) and 144 landmarks (48 x, 48 y, and 48 z)}
}}
\source{
\url{http://datadryad.org/resource/doi:10.5061/dryad.nr210}
}
\usage{
coords.raw
}
\description{
A dataset containing the 48 3D landmarks collected from the skulls of 8 Triturius specimens
and 1 Calotriton specimen (see Ivanovic A, Arntzen JW for more information).
This data set only includes the column regarding the species names and the 114 columns
with the landmarks (48 landmarks x 3 coordinates). You can follow
a detailed tutorial on how to perform a Procrustes analysis with this data set
here: \url{https://sabifo4.github.io/blog/Morphometrics_and_Procrustes_alignment}
}
\keyword{datasets}
