% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBD.R
\name{dBD}
\alias{dBD}
\title{Birth-death process with species sampling}
\usage{
dBD(x, lambda, mu, rho, t1 = 1)
}
\arguments{
\item{x}{numeric, x values for which the density will be plotted. This 
`dbd` function is passed to R function `curve` via variable `expr`. 
The values you type as arguments for variables `from` and `to` in
function `curve` will be used by variable `x` in this function 
(see examples for details).}

\item{lambda}{numeric, birth rate.}

\item{mu}{numeric, death rate.}

\item{rho}{numeric, proportion of species sampled.}

\item{t1}{numeric, age of the phylogeny's root.}
}
\value{
A numeric vector of probability densities.
}
\description{
Kernel density function for the birth-death process with species sampling.
}
\details{
MCMCtree uses the BD kernel to generate the prior on node ages for those
nodes without fossil calibrations. You can look at the examples below for
some suggestions. Note rho must be between 0 and 1. The special case mu =
lambda, rho=0 gives a uniform density. See Yang and Rannala (2006) for full
details.
}
\examples{
# Reproduce Fig. 10.10 from Yang (2014)
# (a) lambda = mu = 1, rho = 0 (uniform density):
curve(dBD(x, 1, 1, 0), xlim=c(0, 1), ylim=c(0, 4), xaxs="i", yaxs="i")

# (b) lambda = 10, mu = 5, rho = 0.01 (old node ages, useful for diversified 
# sampling):
curve(dBD(x, 10, 5, .01), from=0, to=1, lty=2, add=TRUE)

# (c) lambda = 10, mu = 5, rho = 0.001 (old node ages, useful for diversified 
# sampling):
curve(dBD(x, 10, 5, .001), from=0, to=1, lty=3, add=TRUE)

# (d) lambda = 10, mu = 5, rho = 0.99 (young node ages, useful for dense
# sampling of diverse phylogenies):
curve(dBD(x, 10, 5, .99), from=0, to=1, lty=4, add=TRUE)

}
\references{
Yang and Rannala. (2006) Bayesian Estimation of Species Divergence Times
Under a Molecular Clock Using Multiple Fossil Calibrations with Soft Bounds.
\emph{Mol. Biol. Evol.}, 23: 212–226.

Yang (2014) Molecular Evolution: A Statistical Approach. \emph{Oxford 
University Press}
}
\author{
Mario dos Reis
}
