% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmctree.R
\name{gauss.quad}
\alias{gauss.quad}
\title{Estimate marginal likelihood by thermodynamic integration}
\usage{
gauss.quad(mcmcf = "mcmc.txt", betaf = "beta.txt", se = TRUE)
}
\arguments{
\item{mcmcf}{character, mcmc output file name}

\item{betaf}{character, file with beta values}

\item{se}{logical, whether to return the standard error of the estimate}
}
\value{
If \code{se = TRUE}, a numeric vector of size two with the log-marginal
likelihood estimate and the S.E of the estimate. Otherwise just the log-marginal
likelihood.
}
\description{
Estimate marginal likelihood by thermodynamic integration and Gauss-Legendre
quadrature from a sample of \code{n} power posterior MCMC chains sampled
with mcmctree (or bpp).
}
\details{
The MCMC samples should be stored in a directory structure created
by \code{make.bfctlf} with \code{method = "gauss-quad"}. The function 
will read the stored log-likelihood values and calculate the log-marginal 
likelihood. If \code{se = TRUE}, then the standard error of the estimated 
is also given (this requires the \code{coda} package). 

Numerical integration is done using Gauss-Legendre quadrature. See Rannala
and Yang (2017) for details (also dos Reis et al. 2017, Appendix 2).
}
\references{
Rannala B and Yang Z. (2017) Efficient Bayesian species tree inference under 
the multispecies coalescent. \emph{Systematic Biology} 66: 823-842.

dos Reis et al. (2017) Using phylogenomic data to explore the effects of 
relaxed clocks and calibration strategies on divergence time estimation: 
Primates as a test case. \emph{bioRxiv}
}
\seealso{
\code{\link{make.bfctlf}} to prepare directories and mcmctree control files
to calculate the power posterior.
}
\author{
Mario dos Reis
}
