% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{lmk_imp}
\alias{lmk_imp}
\title{Import various landmark files for different specimens at once}
\usage{
lmk_imp(path = NULL, lmk.names = FALSE)
}
\arguments{
\item{path}{Character, absolute path to the directory with the
csv files with the landmark points are.}

\item{lmk.names}{Logical, TRUE if there is an extra column for
landmark names, FALSE otherwise (see details).}
}
\description{
Import more than one csv file with landmark points and return an array
object with dimensions p x k x n, being \code{p} the number of
landmarks, \code{k} the dimension of the coordinates (2D or 3D), and
\code{n} the number of specimens (the number of files, as each file
contains the landmarks for one specimen).
}
\details{
Note that all files need to be comma separated files (csv).

If \code{lmk.names = TRUE}, the format expected for 3D landmarks is
the following:
\tabular{cccccccc}{
landmarks \tab x      \tab y      \tab z      \cr
lmk_1     \tab 0.143  \tab -0.028 \tab -0.030 \cr
lmk_2     \tab 0.128  \tab -0.024 \tab -0.035 \cr
...       \tab ...    \tab ...
}

Otherwise, if \code{lmk.names = TRUE}, then the format is:
\tabular{cccccccc}{
x      \tab y      \tab z      \cr
0.143  \tab -0.028 \tab -0.030 \cr
0.128  \tab -0.024 \tab -0.035 \cr
...    \tab ...    \tab ...
}

Note that you can always have 2D landmarks, so the format is the same
but the column with the \code{z} landmarks will not appear in the files.
}
\author{
Sandra Alvarez-Carretero
}
