% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnL.app.R
\name{lnL.app}
\alias{lnL.app}
\title{Calculate the approximate log-likelihood}
\usage{
lnL.app(
  b,
  b.mle,
  g.mle,
  H.mle,
  transform = c("NT", "SQRT", "ARCSIN"),
  type = c("nt", "aa", "co")
)
}
\arguments{
\item{b}{numeric, the branch length}

\item{b.mle}{numeric, the branch length MLE}

\item{g.mle}{numeric, the gradient at the branch length MLE}

\item{H.mle}{numeric, the second derivative at the branch length MLE}

\item{transform}{character, the transform to be used (see details)}

\item{type}{character, the data type}
}
\description{
Calculate the approximate log-likelihood
}
\details{
Calculates the approximate log-likelihood for two species using
  Taylor expansion of the log-likelihood around the maximum likelihood
  estimate (MLE) of the branch length, using the branch length transforms
  described in dos Reis and Yang (2011).

  Note data type is required for correct calculation of the ARCSIN transform.
  Data types are "nt" for nucleotide, "aa" for amino acid, and "co" for
  codon.
}
\examples{
# This reproduces Fig. 1 in dos Reis and Yang (2011):

# Exact likelihood for two-species function under the JC69 model:
JC69.lnL <- function(b, x, n) {
  p <- .75 - .75 * exp(-4 * b / 3)
  x * log(p) + (n - x) * log (1 - p)
}

# Exact gradient under the JC69 model:
JC69.g <- function(b, x, n) {
  p <- .75 - .75 * exp(-4 * b / 3)
  (x / p - (n - x) / (1 - p)) * exp(-4 * b / 3)
}

# Exact Hessian under the JC69 model:
JC69.H <- function(b, x, n) {
  p <- .75 - .75 * exp(-4 * b / 3)
  (-x/p^2 - (n-x)/(1-p)^2) * exp(-8 * b/3) - 4/3 * (x/p - (n-x)/(1-p)) * exp(-4 * b/3)
} 

# MLE of branch length under the JC69.modle:
JC69.b <- function(x, n) {
  -.75 * log(1 - 4/3 * x/n)
}

par(mfrow=c(1,3))
# Case 1: MLE of b is zero
x. <- 0; n <- 100; 
b <- JC69.b(x., n); lnL.max <- x. * log(1e-6/n) + (n - 1e-6) * log(1 - 1e-6/n)
curve(JC69.lnL(x, x., n) - lnL.max, from=0, to=.6, ylim=c(-50, 0), xlab="b", ylab="lnL", las=1)
g <- JC69.g(1e-6, x., n); H <- JC69.H(1e-6, x., n)
curve(lnL.app(x, 0, g, H, "NT"), lty=2, add=TRUE)
curve(lnL.app(x, 0, g, H, "SQRT"), lty=2, add=TRUE)
curve(lnL.app(x, b, g, H, "ARCSIN"), add=TRUE, lty=3)

# Case 2: MLE of b is between zero and infinity
x. <- 37; n <- 100
b <- JC69.b(x., n); lnL.max <- x. * log(x./n) + (n - x.) * log(1 - x./n)
g <- JC69.g(b, x., n); H <- JC69.H(b, x., n)
curve(JC69.lnL(x, x., n) - lnL.max, from=0, to=2, ylim=c(-50, 0), xlab="b", ylab="lnL", las=1)
curve(lnL.app(x, b, g, H, "NT"), lty=2, add=TRUE)
curve(lnL.app(x, b, g, H, "SQRT"), lty=2, add=TRUE)
curve(lnL.app(x, b, g, H, "ARCSIN"), add=TRUE, lty=3)
abline(v=b, lty=3)

# Case 3: MLE of b is very large
x. <- 74; n <- 100; 
b <- JC69.b(x., n); lnL.max <- x. * log(x./n) + (n - x.) * log(1 - x./n)
curve(JC69.lnL(x, x., n) - lnL.max, from=0, to=4, ylim=c(-50, 0), xlab="b", ylab="lnL", las=1)
g <- JC69.g(b, x., n); H <- JC69.H(b, x., n)
curve(lnL.app(x, b, g, H, "NT"), lty=2, add=TRUE)
curve(lnL.app(x, b, g, H, "SQRT"), lty=2, add=TRUE)
curve(lnL.app(x, b, g, H, "ARCSIN"), add=TRUE, lty=3)
abline(v=b, lty=3)

}
\references{
dos Reis and Yang (2011) Approximate likelihood calculation on a
phylogeny for Bayesian estimation of divergence times. \emph{Molecular
Biology and Evolution}, 28: 2161–2172.
}
