% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-lhd.R
\name{logL.boot}
\alias{logL.boot}
\alias{stepping.stones.boot}
\alias{gauss.quad.boot}
\title{Estimate marginal likelihood from bootstrap replicates}
\usage{
stepping.stones.boot(R, betaf = "beta.txt", preff = "lnL")

gauss.quad.boot(R, betaf = "beta.txt", preff = "lnL")
}
\arguments{
\item{R}{numeric, number of bootstrap replicates used}

\item{betaf}{character, file with beta values}

\item{preff}{character, prefix for files storing boot replicates}
}
\value{
A list with components \code{logml}, the original log-marginal likelihood
estimate, \code{logmlR}, the vector of log-marginal likelihood estimates on
the boostrap replicates, \code{se} and \code{ci}, the standard error and 95\%
credibility interval of \code{logml} calculated on the bootstrap replicates,
and \code{b}, the beta values used.
}
\description{
Estimate marginal likelihood from bootstrap replicates
}
\details{
\code{stepping.stones.boot} and \code{gauss.quad.boot} are used to calculate
the marginal likelihoods on bootstrap replicates using the stepping stones 
and gaussian quadrature methods respectively. The replicates must have
been generated using \link{block.boot}.
}
\seealso{
\link{block.boot}, \link{stepping.stones} and \link{gauss.quad}.
}
