% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{matrix2array}
\alias{matrix2array}
\title{Convert a matrix into a 3D array}
\usage{
matrix2array(X, coords = c(2, 3))
}
\arguments{
\item{X}{Matrix (s x n), 'n' landmark points for 's' specimens
(see details).}

\item{coords}{Integer, 2 or 3 for 2D or 3D landmarks, respectively.}
}
\value{
An object of class array with format k x n x s, where 'k' is the number of
landmarks, 'n' the number of coordinates, and 's' the number of specimens.

Note that if the matrix provided does not have rownames, the specimens in the
returned array (dimension 's') will be labelled as '1', '2', and so on.
See \code{canids19x29.array} in the \code{data} directory for an example of
the format of the object that is returned and
\code{/morpho/data-raw/canids19x29.array.R} for the description of how to
obtain this object from the raw data.
}
\description{
Convert a matrix with landmark points into a 3D array object.
}
\details{
The matrix has format s x n, with 's' rows, one for specimen, and 'n'
columns, one for each coordinate of the landmarks.
Each landmark can be given in 2D or 3D. For instance,
if the landmarks are 3D, the first 3 columns will be the
coordinates x, y, and z for the first landmark, the next 3
columns for the second landmark, and so on:
\tabular{cccccccc}{
 specimens \tab lmk1.x \tab lmk1.y  \tab lmk1.z  \tab lmk2.x \tab lmk2.y \tab lmk2.z  \tab ... \cr
 Sp_1      \tab 0.143  \tab -0.028  \tab -0.044  \tab 0.129  \tab 0.028  \tab -0.043  \tab ... \cr
 Sp_2      \tab 0.128  \tab -0.024  \tab -0.028  \tab 0.124  \tab 0.027  \tab -0.025  \tab ... \cr
 ...       \tab ...    \tab ...     \tab ...     \tab ...    \tab ...    \tab ...     \tab ...
}

See \code{canids19x29.matrix} in the \code{data} directory for an example
and \code{morpho/data-raw/canids19x29.matrix.R} to see how this object
was generated from the raw file.
}
\seealso{
\code{\link{array2matrix}}, \code{\link{write.morpho}}
}
\author{
Sandra Alvarez-Carretero and Mario dos Reis
}
