% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.sum.R
\name{mcmc.sum}
\alias{mcmc.sum}
\title{Calculate summaries from an MCMC run}
\usage{
mcmc.sum(mcmc)
}
\arguments{
\item{mcmc}{a data frame with MCMCtree's output}
}
\value{
A data frame with the mean, median, and 95% CI of times
}
\description{
Calculate summaries from an MCMC run
}
\details{
The data frame should have headers and should contain the output from the 
mcmc.txt file generated by MCMCtree.
}
