% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc2multiphylo.R
\name{mcmc2multiphylo}
\alias{mcmc2multiphylo}
\title{Convert an MCMC sample from MCMCTree or BPP to a list of trees}
\usage{
mcmc2multiphylo(tree, mcmc, time.name, thin)
}
\arguments{
\item{tree}{an object of class phylo}

\item{mcmc}{data frame with an MCMC sample from MCMCTree or a BPP A00
analysis}

\item{time.name}{character vector of length one, this is usually
\code{"tau_"} for BPP, or \code{"t_"} for MCMCTree.}

\item{thin}{numeric, the fraction of MCMC samples to keep}
}
\value{
An object of class multiPhylo (i.e., a list of trees).
}
\description{
Convert an MCMC sample from MCMCTree or BPP to a list of trees
}
\details{
\code{tree} must be rooted and strictly bifurcating, and it must
  match the tree used by BPP or MCMCTree to obtain the MCMC sample. The
  function uses the node ages in \code{mcmc} to calculate branch lengths and
  generate a list of trees (with the same topology as \code{tree}), one tree
  per (thinned) MCMC sample. The tips of the phylogeny are assumed to have
  age zero.
}
\examples{
# hominid MCMC sample from BPP:
data(hominids)
hominid.trees <- mcmc2multiphylo(hominids$tree, hominids$mcmc, "tau_", 0.001)

\dontrun{
# If you have the ape package installed, you can output the trees in 
# Newick format
ape::write.tree(hominid.trees) 
}

}
\author{
Mario dos Reis
}
