% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{proc2MCMCtree}
\alias{proc2MCMCtree}
\title{Procrustes alignment output in MCMCtree format for Bayesian inference}
\usage{
proc2MCMCtree(
  data,
  popdata,
  sp.data,
  sp.popdata,
  filename,
  coords,
  method = c("eigen", "chol"),
  ages,
  ...
)
}
\arguments{
\item{data}{Matrix or array, object with the data set with one specimen per
species. See details.}

\item{popdata}{Matrix or array, containing the specimens for the population of
species with a larger variation (more than one specimen for this species).
See details.}

\item{sp.data}{Numeric, position of the specimen in the \code{data} object
also present in the \code{popdata} object (see details).}

\item{sp.popdata}{Numeric, position of the specimen in the \code{popdata} object
also present in the \code{data} object (see details).}

\item{filename}{Character, name for the output file.}

\item{coords}{Numeric, 2 or 3 for 2D or 3D landmarks, respectively.}

\item{method}{(Optional) character, either \code{"eigen"} or
\code{"col"}, method used to decompose the inverse of the shrinkage
correlation matrix. See details.}

\item{ages}{(Optional) list or vector, ages of the species included in the
morpholical alignment.}

\item{...}{(Optional) Further arguments passed to \code{\link[Morpho]{procSym}}
(see details)}
}
\value{
\item{$dataPS}{Object of class \code{symproc} output by the function
\code{Morpho::procSym}. The user can access the rest of the items in this object
by checking <your_object>$dataPS$<available_Morpho_objects>. It contains the
array with the morphological alignment generated by this function, which
corresponds to the specimens in the object passed to \code{data}. This can
be accessed by loading <your_object>$dataPS$rotated.}
\item{$popdataPS}{Array with the morphological alignment with the object
passed to \code{popdata}, i.e. the alignment with the population sample}
\item{$M}{Matrix with the morphological alignment with the specimens of all
species. Note that this alignment is not corrected for character correlation
nor population noise. The corrected alignment is only printed in the output alignment file.}
\item{$Rsh}{Estimated shrinkage correlation matrix.}
\item{$c}{Estimated population variance.}
}
\description{
Perform a Procrustes alignment with the \code{\link[Morpho]{procSym}}
and \code{\link[Morpho]{align2procSym}} and generate a morphological
alignment in MCMCtree format.
}
\details{
If the objects \code{data} and \code{popdata} are of class "matrix", then they have
\code{s} or \code{ps} rows, respectively, and \code{n} columns as of the amount of characters.
These data sets are supposed to contain landmarks, which can be given in 2D or 3D.
For instance, if the landmarks are 3D, the first 3 columns will be the
coordinates x, y, and z for the first landmark, the next 3
columns for the second landmark; and so on.
\tabular{cccccccc}{
 specimens \tab lmk1.x \tab lmk1.y  \tab lmk1.z  \tab lmk2.x \tab lmk2.y \tab lmk2.z  \tab ... \cr
 Sp_1      \tab 0.143  \tab -0.028  \tab -0.044  \tab 0.129  \tab 0.028  \tab -0.043  \tab ... \cr
 Sp_2      \tab 0.128  \tab -0.024  \tab -0.028  \tab 0.124  \tab 0.027  \tab -0.025  \tab ... \cr
 ...       \tab ...    \tab ...     \tab ...     \tab ...    \tab ...    \tab ...     \tab ...
}
You can load the object \code{C.mat.unal} to see an example of this format and also
take a look at the \code{data-raw/C.R} file for details about how to
generate it. Otherwise, if the objects \code{data} and \code{popdata} are of
class "array", they have format \code{k x q x s} or \code{k x q x ps}, respectively,
where \code{k} is the number of landmarks, \code{q} the number of coordinates,
\code{s} the number of specimens for object \code{data}, and \code{ps} the number of
specimens in a sampled population of one species for object \code{popdata}.
You can load the object \code{C.arr.unal} to see an example of this array format and
also take a look at the \code{data-raw/C.R} file for details about how to
generate it.

The names of the specimens will be taken from either the row names of the data
matrices, if \code{data} and \code{popdata} objects are of class "matrix", or from
the names of the third dimension of the array, if these objects are of class
"array". If no names are found, the specimens will be labelled as "Specimen_1",
"Specimen_2", and so on.

Note that you are providing a population matrix with the landmarks collected from
more than on specimen belonging to the same species. Therefore, it is assumed that
population noise is accounted for.
Furthermore, the landmarks of one of these specimens are also present in the
morphological alignment. First, a Procrusts alignment is performed with the data set
with one specimen per species (\code{data}), and then all the specimens in \code{popdata}
except for the specimen common in \code{data}, which position is specified in the
argument \code{sp.popdata}, are aligned to the mean shape of the
PA generated with \code{data}. Take a look at \code{data.raw/V.R} for more details.

The logarithm of the determinant of the correlation matrix is going to be printed
in the output file to later be used by MCMCtree during the likelihood calculation.

The function \code{\link[Morpho]{procSym}} can perform a Procrustes superimposition
alignment given a dataset of landmarks in array format and allows the user to pass
different options to the arguments defined.
The current function runs with the default parameters
in \code{\link[Morpho]{procSym}} but allows the user to pass three arguments to
\code{\link{proc2MCMCtree}} through \code{...}: scale, reflect, and pairedLM.
If you are thinking of using these arguments, read the documentation in \code{\link[Morpho]{procSym}},
otherwise do not pass further arguments to \code{\link{proc2MCMCtree}} and let it run
\code{\link[Morpho]{procSym}} in default mode.
}
\examples{

# A. Use the unaligned, but processed, carnivoran data (data = C.mat.unal) and
#    vulpes data (popdata = V.mat.unal) to obtain a morphological alignment.
#    The fox specimen that is common in V.mat.unal and C.mat.unal is the
#    one in the first row of V.mat.unal (sp.popdata = 1) and the one in position
#    13 in C.mat.unal (sp.data = 13). The method to
#    decompose the estimate shrinkage correlation matrix (internally calculated
#    in this function) is the Cholesky decomposition (method = c("chol")).
#    We do not add ages.
     right    <- c( 11, 22, 13, 19, 15, 20, 24, 5, 7, 2, 9, 26, 29 )
     left     <- c( 10, 21, 12, 17, 14, 18, 23, 4, 6, 1, 8, 25, 28 )
     pairedLM <- cbind( left, right )
     obj.aln <- proc2MCMCtree( data = C.mat.unal, popdata = V.mat.unal, sp.data = 13,
                               sp.popdata = 1, filename = "./seqfile.aln", coords = 3,
                               method = c("chol"), pairedLM = pairedLM )

# B. Use the unaligned, but processed, carnivoran data (data = C.mat.unal) and
#    vulpes data (popdata = V.mat.unal) to obtain a morphological alignment.
#    The fox specimen that is common in V.mat.unal and C.mat.unal is the
#    one in the first row of V.mat.unal (sp.popdata = 1) and the one in position
#    13 in C.mat.unal (sp.data = 13). The method to
#    decompose the estimate shrinkage correlation matrix (internally calculated
#    in this function) is the Cholesky decomposition (method = c("chol")).
#    We add ages.
     ages <- list( sp1  = 13.135, sp2  =  0.0285, sp3  = 11.95,  sp4  = 35.55,
                   sp5  = 25.615, sp6  = 14.785,  sp7  = 28.55,  sp8  =  0,
                   sp9  =  0,     sp10 =  0,      sp11 =  0,     sp12 =  0,
                   sp13 =  0,     sp14 =  0,      sp15 =  0,     sp16 =  0,
                   sp17 =  6.65,  sp18 =  0.0285, sp19 =  0 )
     right    <- c( 11, 22, 13, 19, 15, 20, 24, 5, 7, 2, 9, 26, 29 )
     left     <- c( 10, 21, 12, 17, 14, 18, 23, 4, 6, 1, 8, 25, 28 )
     pairedLM <- cbind( left, right )
     obj.aln <- proc2MCMCtree( data = C.mat.unal, popdata = V.mat.unal, sp.data = 13,
                               sp.popdata = 1, filename = "./seqfile.aln", coords = 3,
                               method = c("chol"), pairedLM = pairedLM, ages = ages )

}
\seealso{
\code{\link{write.morpho}}, \code{\link{matrix2array}}, \code{\link{array2matrix}}
}
\author{
Sandra Alvarez-Carretero and Mario dos Reis
}
