% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\docType{data}
\name{R}
\alias{R}
\title{Shrunk correlation matrix estimate}
\format{A matrix with n = 21 rows (specimens) and p = 87 columns
(coordinates, 87/3 = 29 landmarks):
\describe{
  \item{n}{Number of specimens belonging to the same species
  (Vulpes vulpes), 21}
  \item{p}{Number of coordinates for which the correlation values
  have been calculated, 87}
}}
\usage{
R
}
\description{
Shrunk estimate of the correlation matrix calculated with the
corpcor::cor.shrink function. The shrinkage parameter of this matrix
is d = 0.6352.
}
\keyword{datasets}
