% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.shrinkage.R
\docType{data}
\name{R.sh}
\alias{R.sh}
\title{Estimated shrinkage correlation matrix}
\format{A matrix of size n x n, where n = 87
(morphological traits, 29 landmarks x 3D coordinates):
\describe{
  \item{n}{Number of traits for which the correlation values
  have been calculated, 87}
}}
\usage{
R.sh
}
\description{
Estimated shrinkage correlation matrix obtained after using the
\code{corpcor::cor.shrink} package.
}
\keyword{datasets}
