% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.shrunk.R
\docType{data}
\name{R.shrunk}
\alias{R.shrunk}
\title{Shrunk correlation matrix estimate}
\format{A matrix of size n x n, where n = 87
(morphological traits, the 87 coordinates):
\describe{
  \item{n}{Number of coordinates for which the correlation values
  have been calculated, 87}
}}
\usage{
R.shrunk
}
\description{
Shrunk estimate of the correlation matrix obtained after applying
the equation \eqn{{R}^{*} = \delta I + (1-\delta){R'}}{R.sh = delta I +
(1-delta)R.unb}
The shrinkage parameter used to estimate the matrix is
is delta = 0.01 so the resulting estimate is closer to the
unbiased correlation matrix.
}
\keyword{datasets}
