% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.unb.R
\docType{data}
\name{R.unb}
\alias{R.unb}
\title{Unbiased correlation matrix estimate}
\format{A matrix of size n x n, where n = 87
(morphological traits, the 87 coordinates):
\describe{
  \item{n}{Number of coordinates for which the correlation values
  have been calculated, 87}
}}
\usage{
R.unb
}
\description{
Unbiased estimate of the correlation matrix obtained for the
population sample of 21 Vulpes vulpes specimens
}
\keyword{datasets}
