% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{sim.pop}
\alias{sim.pop}
\title{Simulate a population matrix}
\usage{
sim.pop(psample, n, c, R)
}
\arguments{
\item{psample}{Numeric, number of specimens the simulated
population sample should include.}

\item{n}{Numeric, number of morphological traits to be simulated.}

\item{c}{Numeric, vector with the population variances with
which the population should be simulated (see details).}

\item{R}{(Optional) matrix, correlation matrix.
(see details).}
}
\value{
\item{P}{Matrix with the simulated population sample of
\code{psample} specimens and \code{n} morphological continuous traits
sampled from a normal distribution with mean = 0 and sd = sqrt(c)}
}
\description{
Simulate a population sample and return in a matrix of
size \code{s x n}, 's' specimens and 'n' characters.
}
\details{
The parameter \code{c} is the population variance and it is used to sample
\code{n} characters for each of the \code{psample} specimens from a
normal distribution \code{x ~ N(0,c)}.
If the population variance is assumed to be the same for all traits,
then the length of \code{c} is 1 and equals to the value of this variance.
If it differs from trait to trait, then a vector of length \code{n} has to
be provided specifying the variance for each of the traits.

If a correlation matrix, \code{R}, is provided, then it is added to the
noisy matrix. Note that the correlation matrix needs to be of class "matrix"
and symmetric.
}
\examples{

# A) Simulation setup: Simulate a population with a sample of
#    psample = 20 specimens, with n = 100 characters, and
#    a low population variance, c = 0.25.

     morpho::sim.pop( psample = 20, n = 100, c = 0.25 )

# B) Simulation setup: Simulate a population with a sample of
#    psample = 20 specimens, with n = 100 characters,
#    a low population variance, c = 0.25, and a low correlation
#    rho = 0.50 used to generate a correlation matrix that follows
#    the constant correlation model (all non-diagonal values
#    equal to rho). This matrix already exists as an object
#    within this package, sim.R, so take a look to
#    morpho/data-raw/sim.R.R to see how it was generated

     morpho::sim.pop( psample = 20, n = 100, c = 0.25, R = sim.R )

}
\seealso{
\code{\link{sim.morpho}}, \code{\link{write.morpho}}
}
\author{
Sandra Alvarez-Carretero and Mario dos Reis
}
