% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.population.R
\docType{data}
\name{sim.population}
\alias{sim.population}
\title{Simulated population}
\format{A matrix of size n x n, where n = 100
(morphological traits, the 100 simulated traits):
\describe{
  \item{n}{Number of simulated traits for which the correlation values
  have been calculated, 100}
}}
\usage{
sim.population
}
\description{
Simulated population with a sample of s = 20 specimens
}
\keyword{datasets}
