% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R.R
\docType{data}
\name{sim.R}
\alias{sim.R}
\title{Correlation matrix for simulations}
\format{A matrix of size n x n, where n = 100
(morphological traits, the 100 simulated traits):
\describe{
  \item{n}{Number of simulated traits for which the correlation values
  have been calculated, 100}
}}
\usage{
sim.R
}
\description{
True correlation matrix used for the examples in the simulations.
The matrix follows the constant correlation model, hence all values
outside the diagonal are rho = 0.50. The size is pxp, being n = 100
the number of characters.
}
\keyword{datasets}
