% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R.shrunk.R
\docType{data}
\name{sim.R.shrunk}
\alias{sim.R.shrunk}
\title{Estimate of the shrunk correlation matrix for a population}
\format{A matrix of size n x n, where n = 100
(morphological traits, the 100 simulated traits):
\describe{
  \item{n}{Number simulated traits for which the correlation values
  have been calculated, 100}
}}
\usage{
sim.R.shrunk
}
\description{
This estimate has been calculated from the previously
generated object "sim.population" as the population
sample (20 specimens).
We used the function calc.pop.cor to generate it and selected
the item "R.shrunk" from the list returned.
}
\keyword{datasets}
