% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-lhd.R
\name{stepping.stones.boot}
\alias{stepping.stones.boot}
\title{Estimate marginal likelihood by stepping stones on bootstrap replicates}
\usage{
stepping.stones.boot(R, betaf = "beta.txt", preff = "lnL")
}
\arguments{
\item{R}{numeric, number of bootstrap replicates used}

\item{betaf}{character, file with beta values}

\item{preff}{character, prefix for files storing boot replicates}
}
\value{
A list with components \code{logml}, the original log-marginal likelihood
estimate, \code{logmlR}, the vector of log-marginal likelihood estimates on
the boostrap replicates, \code{se} and \code{ci}, the standard error and 95\%
credibility interval of \code{logml} calculated on the bootstrap replicates,
and \code{b}, the beta values used.
}
\description{
Estimate marginal likelihood by stepping stones on bootstrap replicates
}
\details{
The stepping stones method is used to calculate the marginal likelihoods on
bootstrap replicates. The replicates must have been generated using  
\link{block.boot}.
}
\seealso{
\link{block.boot} and \link{stepping.stones}.
}
