% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V.arr.unal.R
\docType{data}
\name{V.arr.unal}
\alias{V.arr.unal}
\title{29 3D landmarks from the skulls of 21 Vulpes vulpes specimens before Procrustes analysis}
\format{
An array with k = 29 (landmarks), q = 3 (coordinates) and s = 21 (specimens):
\describe{
  \item{k}{landmark points collected from 21 foxes specimens, 29}
  \item{q}{coordinates in 3D or 2D, 3}
  \item{s}{number of specimens, 21}
}
}
\usage{
V.arr.unal
}
\description{
A 3D array containing the 29 3D landmarks collected from the skulls of 21 "Vulpes vulpes"
specimens before carrying out a Procrustes analysis (PA).
Please take a look at the description in morpho/data-raw/V.R to understand
how this object was generated.
}
\keyword{datasets}
