% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V.PS.nov1.R
\docType{data}
\name{V.PS.nov1}
\alias{V.PS.nov1}
\title{Object of class array output by Morpho after PA}
\format{
Object procSym
  \describe{
  \item{...}{Check \code{\link[Morpho]{procSym}} for more details}
}
}
\usage{
V.PS.nov1
}
\description{
Object of class array output by Morpho after being aligned to the
mean shape of the 19 carnivoran species previously generated (object "C.PS").
Please take a look at the description in morpho/data-raw/V.R to understand
how this object was generated.
}
\keyword{datasets}
