% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V.R
\docType{data}
\name{V}
\alias{V}
\title{29 3D landmarks from the skulls of 19 carnivoran specimens after Procrustes analysis}
\format{
A matrix with s = 21 rows and n = 87 columns (87/3 = 29 landmarks):
  \describe{
  \item{s}{Rows, specimens from which landmarks were collected, 21}
  \item{n}{Columns, 87 traits (29 landmarks in 3D) after the PA}
}
}
\usage{
V
}
\description{
A matrix containing the 29 3D landmarks collected from the skulls of 21 "Vulpes vulpes"
specimens after carrying out a Procrustes analysis (PA).
Please take a look at the description in morpho/data-raw/V.R to understand
how this object was generated.
}
\keyword{datasets}
