% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.foxes.R
\docType{data}
\name{var.foxes}
\alias{var.foxes}
\title{Vector with the population variance of Vulpes vulpes}
\format{A vector with i = 87 variances regarding the "Vulpes vulpes"
population:
\describe{
  \item{i}{Number of variances for the foxes population, 87}
}}
\usage{
var.foxes
}
\description{
Vector with 87 within-species variances calculated from the object V.
Please take a look at the description in morpho/data-raw/var.foxes.R to understand
how this object was generated.
}
\keyword{datasets}
