% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.fx.R
\docType{data}
\name{var.fx}
\alias{var.fx}
\title{Vector with the population variance of Vulpes vulpes}
\format{A vector with i = 87 variances regarding the Vulpes vulpes
population:
\describe{
  \item{i}{Number of population variances regarding the 21 Vulpes vulpes
  specimens, 87}
}}
\usage{
var.fx
}
\description{
Vector with 87 variances calculated from the matrix with the
21x87 morphological characters (Vulpes vulpes specimens)
after the Procrustes analysis (PA) was performed.
First of all, the variance-covariance matrix of the Vulpes data set
with the morphological aligned characters is calculated. Afterwards,
the diagonal of this variance-covariance matrix contains the variances
for the Vulpes vulpes population.
}
\keyword{datasets}
