% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulpes21x29.array.R
\docType{data}
\name{vulpes21x29.array}
\alias{vulpes21x29.array}
\title{29 3D landmarks from the skulls of 21 Vulpes vulpes specimens after Procrustes analysis}
\format{An array with k = 29 (landmarks), n = 3 (coordinates) and s = 21 (specimens):
\describe{
  \item{k}{landmark points collected from 21 Vulpes vulpes specimens, 29}
  \item{n}{coordinates in 3D or 2D, 3}
  \item{s}{number of specimens, 21}
}}
\source{
\url{https://github.com/sabifo4/Morphometrics_practical}
}
\usage{
vulpes21x29.array
}
\description{
A 3D array containing the 29 3D landmarks collected from the skulls of 21 Vulpes vulpes
specimens after carrying out a Procrustes analysis (PA). You can follow a detailed
tutorial on how to perform the PA here:
\url{https://sabifo4.github.io/blog/Morphometrics_and_Procrustes_alignment}
}
\keyword{datasets}
