% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{write.morpho}
\alias{write.morpho}
\title{Generate a phylip file for MCMCTree}
\usage{
write.morpho(filename, proc, coords = c(2, 3), names = NULL, ages = NULL,
  popvar = NULL, R = NULL, method = c("eigen", "chol"), ...)
}
\arguments{
\item{filename}{Character, name for the output file.}

\item{proc}{3D array (p x k x n) or matrix (n x p), landmark coordinates
after a Procrustes analysis (see details).}

\item{coords}{Integer, only needed if \code{class(proc) == "array"}
(see details).}

\item{names}{(Optional) list, species name included in the morphological
alignment (see examples B and C).}

\item{ages}{(Optional) list, ages of the species included in the
morpholical alignment (see example C).}

\item{popvar}{(Optional) vector of integers, population variance (see details).}

\item{R}{(Optional) matrix, correlation matrix. Requires \code{popvar} and
\code{method}.}

\item{method}{(Optional) character, method to decompose the inverse of the
correlation matrix R, \code{eigen} or \code{chol} (see details).
Requires \code{popvar} and \code{R}.}

\item{...}{Further arguments passed to \code{write.morpho()} (see details).}
}
\description{
Generate an alignment file in phylip format for MCMCTree.
The option "seqfile" in the control file used by MCMCTree
should read the path to the file output by this function.
}
\details{
If \code{class(proc) == "array"}, the array has format p x k x n, where
'p' is the number of landmarks, 'k' the number of coordinates,
and 'n' the number of specimens. See \code{canids19x29.array} in the
\code{data} directory for an example.
You also need to provide \code{coords = 2} (2D landmark points) or
\code{coords = 3} (3D landmark points) if the \code{proc} object is an array.

If \code{class(proc) == "matrix"}, the matrix has 'n' rows, one for specimen,
and 'p' columns.Each landmark can be given in 2D or 3D. For instance,
if the landmarks are 3D, the first 3 columns will be the
coordinates x, y, and z for the first landmark, the next 3
columns for the second landmark, and so on.
\tabular{cccccccc}{
 specimens \tab lmk1.x \tab lmk1.y  \tab lmk1.z  \tab lmk2.x \tab lmk2.y \tab lmk2.z  \tab ... \cr
 Sp_1      \tab 0.143  \tab -0.028  \tab -0.044  \tab 0.129  \tab 0.028  \tab -0.043  \tab ... \cr
 Sp_2      \tab 0.128  \tab -0.024  \tab -0.028  \tab 0.124  \tab 0.027  \tab -0.025  \tab ... \cr
 ...       \tab ...    \tab ...     \tab ...     \tab ...    \tab ...    \tab ...     \tab ...
}
See \code{canids19x29.matrix} in the \code{data} directory for a
more detailed example.

If \code{names} is not provided, the name for each species will be
"Species_1", "Species_2", and so on.

The parameter \code{popvar} accounts for the population variance. First of all,
a PA regarding the landmark points collected from different specimens
belonging to different species should be carried out, which will result into
a matrix of superimposed coordinates, say \code{M}. If there are many specimens of
one of the species included in \code{M}, then another PA is carried out with
the landmark points collected from these specimens belonging to the same species.
The resulting matrix, say \code{V}, is used to obtain the population variance.
The variance-covariance matrix of \code{V}, say \code{VCV}, is first calculated
and, afterwards, the diagonal of \code{VCV}, the vector of variances, is used
to scale the matrix \code{M} according to the standard deviation.
Therefore, \code{popvar = diag(VCV)}. If this parameter is not provided, then
the population variance is assumed to be 0.

If \code{method = "eigen"}, the inverse of the correlation matrix provided,
 \code{R}, will be decomposed following the spectral eigendecomposition
(see \code{eigen}).
If \code{method = "chol"}, the inverse of \code{R} will be decomposed
following the Cholesky decomposition (see \code{chol}).

There are three further parameters you can pass to \code{write.morpho} through
\code{ ... }: \code{R.sh}, a shrunk matrix, \code{scaled}, a logical value
indicating if the object \code{proc} has been already scaled, and a numeric value
for the maximum age, which is always 1.
These arguments are used within the function \code{\link{simul.morpho}}
and are not thought to be called by the user.
}
\examples{
# A.1) Providing only the morphological alignment (proc) after the
#      Procrustes analysis (PA) in an object of class \\"array\\".

       write.morpho( filename = "seqfile.aln", proc = canids19x29.array,
                     coords = 3 )

# A.2) Providing only the morphological alignment (proc) after the
#      Procrustes analysis (PA) in an object of class \\"matrix\\".

       write.morpho( filename = "seqfile.aln", proc = canids19x29.matrix )

# B) Providing the morphological alignment (proc) after the
#    PA in an object of class \\"array\\" and a list with the
#    names of the species

     names <- list( sp1  = "Ael_sp.", sp2  = "Can_dir", sp3  = "Epi_hay", sp4  = "Hes_sp.",
                    sp5  = "Mes_cor", sp6  = "Tom_sp.", sp7  = "Enh_pah", sp8  = "Cuo_alp",
                    sp9  = "Spe_ven", sp10 = "Can_lup", sp11 = "Cer_tho", sp12 = "Oto_meg",
                    sp13 = "Vul_vul", sp14 = "Urs_ame", sp15 = "Ail_ful", sp16 = "Nan_bio",
                    sp17 = "Par_her", sp18 = "Tha_won", sp19 = "Smi_fat"
                  )

     write.morpho( filename = "seqfile.aln", proc = canids19x29.array,
                   coords = 3, names = names )

# C) Providing the morphological alignment (proc) after the
#    PA in an object of class \\"array\\", a list with the names
#    of the species, and a list with the ages of the species

     names <- list( sp1  = "Ael_sp.", sp2  = "Can_dir", sp3  = "Epi_hay", sp4  = "Hes_sp.",
                    sp5  = "Mes_cor", sp6  = "Tom_sp.", sp7  = "Enh_pah", sp8  = "Cuo_alp",
                    sp9  = "Spe_ven", sp10 = "Can_lup", sp11 = "Cer_tho", sp12 = "Oto_meg",
                    sp13 = "Vul_vul", sp14 = "Urs_ame", sp15 = "Ail_ful", sp16 = "Nan_bio",
                    sp17 = "Par_her", sp18 = "Tha_won", sp19 = "Smi_fat"
                  )

     ages <- list( sp1  = 15.97, sp2  =  1.80, sp3  = 13.6,  sp4  = 39.74,
                   sp5  = 30.80, sp6  = 15.97, sp7  = 30.80, sp8  =  0,
                   sp9  =  0,    sp10 =  0,    sp11 =  0,    sp12 =  0,
                   sp13 =  0,    sp14 =  0,    sp15 =  0,    sp16 =  0,
                   sp17 =  0,    sp18 =  0,    sp19 =  0
                  )

     write.morpho( filename = "seqfile.aln", proc = canids19x29.array,
                   coords = 3, names = names, ages = ages )

# D) Providing an object of class \\"array\\" after having
#    carried out a PA. As an example, we use the function
#    geomorph::gpagen, but you can use your preferred
#    function meanwhile the format of the \\"proc\\" object
#    (class array) for write.morpho is
#    "p landmarks x k coordinates x n species"

     df <- canids19x29.raw

     # Process raw data

     mm <- df[,2:dim( df )[2]] # 19 sp x 87 coords (87/3=29 lmks)
     rownames( mm ) <- df[,1]
     mm <- as.matrix( mm )

     # Convert matrix into 3D array

     ma <- matrix2array( proc = mm, coords = 3 )

     # Get procrustes analysis done

     ma.paln <- geomorph::gpagen( ma )

     # Check object is class \\"array\\"

     class( ma.paln$coords )

     # Run write.morpho

     write.morpho( filename = "seqfile.aln", proc = ma.paln$coords, coords = 3 )

# E) Providing the morphological alignment (proc) after the
#    PA in an object of class \\"matrix\\" and population variance

     write.morpho( filename = "seqfile.aln", proc = canids19x29.array,
                   coords = 3, popvar = var.fx )

# F) Providing the morphological alignment (proc) after the
#    PA in an object of class \\"matrix\\", and the population variance
#    and the correlation matrix (choosing the Cholesky
#    decomposition)

     write.morpho( filename = "seqfile.aln", proc = canids19x29.array,
                   coords = 3, popvar = var.fx, R = R, method = "chol" )




}
\seealso{
\code{\link{matrix2array}}, \code{\link{array2matrix}}, \code{\link{simul.morpho}}
}
\author{
Sandra Alvarez-Carretero
}
