% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{write.phylo}
\alias{write.phylo}
\title{Generate a phylip file for MCMCTree}
\usage{
write.phylo(proc, names = NULL, ages = NULL)
}
\arguments{
\item{proc}{Array ( p landmarks x k coordinates x n species),
matrix (n species x p landmarks), or data.frame (n species x p landmarks)
object with ONLY the landmarks of the bones of the species
that are included in the alignment. Note that the Procrustes
superimposition should have been done before generating
this file.}

\item{names}{(optional) List with the species name included in the
morphological alignment. If not provided, the name for each species
will be "Species_1", "Species_2", and so on.
E.g. sp.list <- list( sp1 = "sp.1", sp2 = "sp.2", ... )}

\item{ages}{(optional) List of the ages of the species included
in the morpholical alignment.
E.g. age.list <- list( sp1 = 15, sp2 = 0, ... )}
}
\value{
Alignment file in phylip format for MCMCTree.
}
\description{
Generate an alignment file in phylip format for MCMCTree.
The option "seqfile" in the control file used by MCMCTree
should read the path to the file output by this function.
}
\examples{

# A) Providing only the morphological alignment (proc)

     data()
     write.phylo( proc = coords.proc )

# B) Providing the morphological alignment (proc) and a list
#    with the names of the species

     data()
     names <- list( sp1 = "Sp.011", sp2 = "Sp.23", sp3 = "Sp.1333", sp4 = "Sp.00127",
                    sp5 = "Sp.23625", sp6 = "Sp.1304", sp7 = "Sp.543", sp8 = "Sp.02",
                    sp9 = "Sp.234" )

     write.phylo( proc = coords.proc, names = names )

# C) Providing the morphological alignment (proc), a list
#    with the names of the species, and a list with the
#    ages of the species

     data()
     names <- list( sp1 = "Sp.011", sp2 = "Sp.23", sp3 = "Sp.1333", sp4 = "Sp.00127",
                    sp5 = "Sp.23625", sp6 = "Sp.1304", sp7 = "Sp.543", sp8 = "Sp.02",
                    sp9 = "Sp.234" )

     ages <- list( sp1 = 15, sp2 = 30, sp3 = 8, sp4 = 0,
                   sp5 = 0, sp6 = 0, sp7 = 12, sp8 = 0,
                   sp9 = 25.5 )

     write.phylo( proc = coords.proc, names = names, ages = ages )

# D) Providing an object of class \\"array\\" after having
#    carried out a Procrustes analysis (PA). As an example,
#    we use the function geomorph::gpagen, but you can use
#    your preferred function meanwhile the format of the
#    \\"proc\\" object for write.phylo is "p landmarks x k coordinates x n species"

     data()
     df <- coords.raw
     mm <- df[,2:( dim( df )[2] )] # 9sp x 144coords (144/3=48 lmk)
     rownames( mm ) <- df[,1]

     # Create an empty array to store the coordinates in the format
     # p x k x n (num.coords x coor3D x ns)

     num.coords <- dim(mm)[2]
     coords <- 3 # Change according to your data set

     ns <- 9
     ma <- array( dim = c( num.coords / coords, coords, ns ) ) # 48 lmks, 3D, 9 specimens
     dimnames( ma ) <- list( paste( "lmk", seq( 1:( num.coords/coords ) ), sep="" ),
                             c("x", "y", "z"),
                             df[,1])

     # Select x, y, z positions

     xi <- seq( from=1, to = num.coords, by = 3)
     yi <- seq( from=2, to = num.coords, by = 3)
     zi <- seq( from = 3, to = num.coords, by = 3)

     # Fill in array p x k x n

     for (i in 1:ns) {
        ma[,1,i] <- unlist( mm[i,xi] )
        ma[,2,i] <- unlist( mm[i,yi] )
        ma[,3,i] <- unlist( mm[i,zi] )
     }

     # Get procrustes analysis done

     ma.paln <- geomorph::gpagen(ma)

     # Check object is class \\"array\\"

     class( ma.paln$coords )

     # Run write.phylo

     write.phylo( proc = ma.paln$coords )

}
\author{
Sandra Alvarez-Carretero
}
