% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canids19x29.array.R
\docType{data}
\name{canids19x29.array}
\alias{canids19x29.array}
\title{29 3D landmarks from the skulls of 19 carnivoran specimens after Procrustes analysis}
\format{An array with p = 29 (landmarks), k = 3 (coordinates) and n = 19 (specimens):
\describe{
  \item{p}{landmark points collected from 19 carnivoran specimens, 29}
  \item{k}{coordinates in 3D or 2D, 3}
  \item{n}{number of specimens, 19}
}}
\source{
\url{https://github.com/sabifo4/Morphometrics_practical}
}
\usage{
canids19x29.array
}
\description{
A 3D array containing the 29 3D landmarks collected from the skulls of 19 carnivoran
specimens after carrying out a Procrustes analysis (PA). You can follow a detailed
tutorial on how to perform the PA here:
\url{https://sabifo4.github.io/blog/Morphometrics_and_Procrustes_alignment}
}
\keyword{datasets}
