% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{matrix2array}
\alias{matrix2array}
\title{Convert a matrix into a 3D array}
\usage{
matrix2array(proc, coords = c(2, 3))
}
\arguments{
\item{proc}{Matrix (n x p), 'p' landmark points for 'n' specimens
(see details).}

\item{coords}{Integer, 2 or 3 for 2D or 3D landmarks, respectively.}
}
\value{
An object of class array with format p x k x n, where 'p' is the number of
landmarks, 'k' the number of coordinates, and 'n' the number of specimens.

Note that if the matrix provided does not have rownames, the specimens in the
returned array (dimension 'n') will be labelled as \"1\", \"2\", and so on.
}
\description{
Convert a matrix with landmark points into a 3D array object.
}
\details{
The matrix has format n x p, with 'n' rows, one for specimen, and 'p' columns.
Each landmark can be given in 2D or 3D. For instance,
if the landmarks are 3D, the first 3 columns will be the
coordinates x, y, and z for the first landmark, the next 3
columns for the second landmark, and so on. See \code{canids19x29.matrix} in the
\code{data} directory for an example.
}
\author{
Sandra Alvarez-Carretero
}
