% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{matrix2array}
\alias{matrix2array}
\title{Convert a matrix into an array}
\usage{
matrix2array(X, coords = c(2, 3))
}
\arguments{
\item{X}{Matrix  of size \code{s x n}, \code{n} landmark points for \code{s} specimens
(see details).}

\item{coords}{Numeric, 2 or 3 for 2D or 3D landmarks, respectively.}
}
\value{
An object of class array with format \code{k x q x s}, where \code{k} is the number of
landmarks, \code{q} the number of coordinates, and \code{s} the number of specimens.

Note that if the matrix provided does not have rownames, the specimens in the
returned array (names for the 's' matrices accessed through the array, i.e.
\code{dimnames( array )[ 3 ]} will be labelled as '1', '2', and so on.
See object \code{C.arr.unal} for an example of the format of the object that is
returned and \code{data-raw/C.R} for the description of how to
obtain this object.
}
\description{
Convert a matrix with landmark points into an object of class array.
}
\details{
The matrix has format \code{s x n}, with \code{s} rows, one per specimen, and \code{n}
columns, one for each coordinate of the landmarks.
Each landmark can be given in 2D or 3D. For instance,
if the landmarks are 3D, the first 3 columns will be the
coordinates x, y, and z for the first landmark, the next 3
columns for the second landmark; and so on:
\tabular{cccccccc}{
 specimens \tab lmk1.x \tab lmk1.y  \tab lmk1.z  \tab lmk2.x \tab lmk2.y \tab lmk2.z  \tab ... \cr
 Sp_1      \tab 0.143  \tab -0.028  \tab -0.044  \tab 0.129  \tab 0.028  \tab -0.043  \tab ... \cr
 Sp_2      \tab 0.128  \tab -0.024  \tab -0.028  \tab 0.124  \tab 0.027  \tab -0.025  \tab ... \cr
 ...       \tab ...    \tab ...     \tab ...     \tab ...    \tab ...    \tab ...     \tab ...
}

See object \code{C} for an example of its format and \code{data-raw/C.R} to
see how this object is generated.
}
\seealso{
\code{\link{array2matrix}}, \code{\link{write.morpho}}
}
\author{
Sandra Alvarez-Carretero and Mario dos Reis
}
