% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancrec.R
\name{mcmc2anc}
\alias{mcmc2anc}
\title{Fast ancestral character reconstruction from an MCMC sample
TODO: Implement covariance transform and do proper example}
\usage{
mcmc2anc(tree, M, mcmc, time.name, rate.name, tip.ages)
}
\description{
Fast ancestral character reconstruction from an MCMC sample
TODO: Implement covariance transform and do proper example
}
\examples{
tt = ape::read.tree("~/phyl/carnivores/tutorial/carnivores.tree")
Mdf <- read.table("~/phyl/carnivores/tutorial/M.txt")
M <- Mdf[,-1]; rownames(M) <- Mdf[,1]
mcmc = read.table("~/phyl/carnivores/tutorial/mcmc.txt", head=TRUE)
back.ages = as.numeric(unlist(strsplit(tt$tip.label, "\\\\^"))[seq(from=2, to=2*19, by=2)])
back.ages = max(back.ages) - back.ages
xx = mcmc2anc(tt, M, mcmc=mcmc, time.name="t_", rate.name="r_g1_", tip.ages=back.ages)
tr <-  ape::read.tree("~/phyl/carnivores/tutorial/rate_g1.tree")
plot(xx$rates, tr$edge.length)
}
