% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{sim.morpho}
\alias{sim.morpho}
\title{Simulate a continuous morphological alignment}
\usage{
sim.morpho(tree, n, c = 0, R, ...)
}
\arguments{
\item{tree}{Phylo, object with a phylogenetic tree
(see \code{\link[ape]{rTraitCont}}).}

\item{n}{Numeric, number of morphological traits to be simulated.}

\item{c}{(Optional) numeric, vector with population variances to
add to the simulated morphological traits (see details).}

\item{R}{(Optional) matrix, correlation matrix (see details).}

\item{...}{Further arguments passed to \code{\link[ape]{rTraitCont}}
(see details).}
}
\value{
\item{M}{Matrix with the simulated morphological continuous data accounting
for noise and, if provided, for population variance and trait correlation too.}
}
\description{
Simulate a continuous morphological alignment using \code{\link[ape]{rTraitCont}}
and later allowing to account for population variance and trait correlation.
}
\details{
The function \code{\link[ape]{rTraitCont}} simulates continuous traits and
can take different parameters to adjust the simulation
(e.g. the model, the rate drift, etc.).
These parameters are the ones the user can pass through
\code{sim.morpho()}.
The default values that \code{sim.morpho()} uses are
\code{model = "BM"}, \code{sigma = 1}, \code{ancestor = F},
and \code{root.value = 0}. For this kind of simulation,
\code{sim.morpho} allows only \code{ancestor = F}, so please
do not change this parameter.
In the \code{\link[ape]{rTraitCont}} package, the parameter
\code{model} can be \code{model = BM}, \code{model = OU}, or
a function \code{model = FUN} provided by the user. Currently,
\code{sim.morpho()} supports only the first two.

The parameter \code{c} is the population variance and used to simulate
the noise matrix, which follows a normal distribution \code{x ~ N(0,c)}.
If the population variance is assumed to be the same for all traits,
then the length of \code{c} is 1 and equals to the value of this variance.
If it differs from trait to trait, then a vector of length \code{n} has to
be provided specifying the variance for each of the traits.

The simulated noise is later added to the morphological data previously
generated, so we obtain the noisy matrix.
If a correlation matrix, \code{R}, is provided, then it is added to the
noisy matrix. Note that the correlation matrix needs to be of class "matrix"
and symmetric.
}
\examples{
# A) Simulation setup: Simulate a morphological alignment
#    with n = 100 continuous characters for a phylogeny
#    defined in object 'tree', with the default parameters in
#    'sim.morpho' to run 'rTraitCont'.
#    Population variance and correlation are not considered,
#    i.e. c = 0 and R not provided.

     morpho::sim.morpho( tree = sim.tree, n = 100 )


# B) Simulation setup: Simulate a morphological alignment
#    with n = 100 continuous characters for a phylogeny
#    defined in object 'tree', but with different parameters
#    than the default ones in 'sim.morpho' to run
#    'rTraitCont'. Population variance and correlation are not
#    considered, i.e. c = 0 and R not provided.

     morpho::sim.morpho( tree  = sim.tree, n = 100,
                         model = "OU", sigma = 0.2, alpha = 2 )


# C) Simulation setup: Simulate a morphological alignment
#    with n = 100 continuous characters for a phylogeny
#    defined in object 'tree', with the default parameters in
#    'sim.morpho' to run 'rTraitCont'.
#    Population variance is low, c = 0.25, but correlation is not
#    considered, i.e. R not provided.

     morpho::sim.morpho( tree = sim.tree, n = 100, c = 0.25 )


# D) Simulation setup: Simulate a morphological alignment
#    with n = 100 continuous characters for a phylogeny
#    defined in object 'tree', with the default parameters in
#    'sim.morpho' to run 'rTraitCont'.
#    Population variance is low, c = 0.25, and a correlation
#    matrix is provided.

     morpho::sim.morpho( tree = sim.tree, n = 100, c = 0.25, R = sim.R )

}
\seealso{
\code{\link{write.morpho}}
}
\author{
Sandra Alvarez-Carretero and Mario dos Reis
}
