% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{sim.pop}
\alias{sim.pop}
\title{Simulate a population matrix}
\usage{
sim.pop(psample, n, c, R)
}
\arguments{
\item{psample}{Numeric, number of specimens the simulated
population sample should include.}

\item{n}{Numeric, number of morphological traits to be simulated.}

\item{c}{Numeric, vector with the variances for the species 
within a population (see details).}

\item{R}{(Optional) matrix, correlation matrix.
(see details).}
}
\value{
\item{$P}{Matrix with the simulated population sample}
\item{$var}{Vector with the estimated variances}
\item{$Rsh}{Estimated shrinkage correlation matrix, only returned if \code{R} is provided}
}
\description{
Simulate a population sample and return a list with (i) a matrix of
size \code{s x n}, \code{s} specimens and \code{n} characters, (ii) a
vector with the estimated population variances for each character,
and (iii) the estimated shrinkage correlation matrix if the true
correlation matrix is provided.
}
\details{
The parameter \code{c} is the population noise and it is used to sample
\code{n} characters for each of the \code{psample} specimens from a
normal distribution \code{x ~ N(0,c)}.
If the population noise is assumed to be the same for all the characters
within the species, then the length of \code{c} is 1 and equals to the value
of this variance.
If it differs, then a vector of length \code{n} has to
be provided specifying the variance for each of the characters

If a correlation matrix, \code{R}, is provided, then it is added to the
population matrix. Note that the correlation matrix needs to be of class "matrix"
and symmetric. You can take a look at \code{data-raw/sim.R.R} to follow
the commands used to generate this matrix, object \code{R.sim}, which is used
in the examples.
}
\examples{

# A) Simulation setup: Simulate a population with
#    psample = 20 specimens, and sample n = 100 characters with
#    a low population noise, c = 0.25.

     sim.pop( psample = 20, n = 100, c = 0.25 )

# B) Simulation setup: Simulate a population with
#    psample = 20 specimens, and sample n = 100 characters with
#    a low population noise, c = 0.25, and a low trait correlation
#    rho = 0.50 (correlation matrix that follows
#    the constant correlation model, i.e. all non-diagonal values
#    equal to rho).

     sim.pop( psample = 20, n = 100, c = 0.25, R = sim.R )

}
\seealso{
\code{\link{sim.morpho}}, \code{\link{write.morpho}}
}
\author{
Sandra Alvarez-Carretero and Mario dos Reis
}
