% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{write.morpho}
\alias{write.morpho}
\title{Generate a file with a morphological alignment for MCMCtree}
\usage{
write.morpho(
  M,
  filename,
  c = 0,
  R = diag(1, dim(M)[2]),
  method,
  A = NULL,
  names,
  ages,
  scale = 1
)
}
\arguments{
\item{M}{Matrix, \code{s} rows (specimens) and \code{n} morphological
continuous characters (see details).}

\item{filename}{Character, name for the output file.}

\item{c}{Numeric, vector of variances within the species of a population (see details).
If not provided, \code{c = 0} (no population noise).}

\item{R}{Matrix, correlation matrix. Requires \code{method}
(see details). If not provided, \code{R = I} (no character correlation).}

\item{method}{(Optional) character, either \code{"eigen"} or
\code{"chol"}, method used to decompose the inverse of the shrinkage
correlation matrix. Requires \code{R} (see details).}

\item{A}{(Optional) matrix, decomposed matrix. Requires \code{R}
but not \code{method} (see details).}

\item{names}{(Optional) list or character, species name included in the
morphological alignment (see examples B and C).}

\item{ages}{(Optional) list or numeric, ages of the species included in the
morphological alignment (see example C).}

\item{scale}{numeric, all characters are multiplied by \code{scale} before
the morphological matrix is printed to file. Useful to re-scale the characters
so they have the required variance (rate).}
}
\description{
Generate an alignment file with quantitative characters in MCMCtree format.
The option "seqfile" in the control file used by MCMCtree
should read the path to the file output by this function.
}
\details{
The matrix \code{M} has \code{s} rows, one for each specimen, and \code{n}
columns regarding the characters. If the data set contains landmarks,
they can be given in 2D or 3D.
For instance, if the landmarks are 3D, the first 3 columns will be the
coordinates x, y, and z for the first landmark; the next 3
columns for the second landmark; and so on.
\tabular{cccccccc}{
 specimens \tab lmk1.x \tab lmk1.y  \tab lmk1.z  \tab lmk2.x \tab lmk2.y \tab lmk2.z  \tab ... \cr
 Sp_1      \tab 0.143  \tab -0.028  \tab -0.044  \tab 0.129  \tab 0.028  \tab -0.043  \tab ... \cr
 Sp_2      \tab 0.128  \tab -0.024  \tab -0.028  \tab 0.124  \tab 0.027  \tab -0.025  \tab ... \cr
 ...       \tab ...    \tab ...     \tab ...     \tab ...    \tab ...    \tab ...     \tab ...
}
See descriptions in \code{data-raw/carnivores19x29.raw.R},
\code{data-raw/vulpes21x29.raw.R}, \code{data-raw/C.R}, \code{data-raw/V.R},
to know how to obtain the morphological alignment used in \code{write.morpho}.
Note that the explanation starts with the processing of raw data.

If the data set contains a set of \code{n} morphological continuous characters,
e.g. from a simulated data set, the file should look like
\tabular{cccccc}{
 specimens \tab char.1  \tab char.2   \tab char.3   \tab char.4  \tab ... \cr
 Sp_1      \tab 0.143   \tab -0.028   \tab -0.044   \tab 0.129   \tab ... \cr
 Sp_2      \tab 0.128   \tab -0.024   \tab -0.028   \tab 0.124   \tab ... \cr
 ...       \tab ...     \tab ...      \tab ...      \tab ...     \tab ...
}

Note that if a list with the specimens names is not passed to the parameter \code{names},
the name for each species will be obtained from the row names in object \code{M}.
Please note that this feature has changed as of April 2025; tags as
"Species_1", "Species_2", and so on would have been given instead. 
Nevertheless, if you are using function \code{sim.pop} to simulate data,
tags that follow notation "Species_X" will be used for specimens names.

The object \code{c} can be of length 1, if all characters have the same variance, or
a vector of length \code{n} with the variance of each of the characters. For the latter,
you can take a look at object \code{var.foxes}, which has been generated following
the steps explained in \code{data-raw/var.foxes.R}.

The object \code{R} has to be a symmetric and positive definite object of class matrix,
i.e. \code{class( R ) = "matrix"}. See \code{R.sh} for an example of its format.
You can also read the description in \code{data-raw/R.shrinkage.R} for the details
about how to generate this matrix.

The logarithm of the determinant of the correlation matrix is going to be printed
in the output file to later be used by MCMCtree during the likelihood calculation.

If a correlation matrix \code{R} is provided, \code{write.morpho} can use either
the \code{method = "chol"} or \code{method = "eigen"} to get a matrix \code{A}
such that \eqn{\mathrm{R^{-1}}=\mathrm{A^{T}}\mathrm{A}}{R^{-1} = t(A) * A}.
This matrix \eqn{\mathrm{A^{T}}}{t(A)} is later used to transform
the morphological data to account for the correlation in this data set,
so that the transformed characters in \code{Z},
\eqn{\mathrm{Z}=\mathrm{M}\mathrm{A^{T}}}{Z = M * t(A)}, are independent.
Alternatively, this matrix \code{A} can also be provided by the user. You can
read the description to generate this matrix in \code{data-raw/A.R} to
understand how to generate this matrix, which is also available as object \code{A}.
If you decide to use a matrix \code{A}, it will be used to transform the data
and no decomposition will be performed, thus saving computational time when
large matrices are to be used.
}
\examples{
# A.1) Providing the morphological alignment (M = C) and
#      the name for the output file. This does not account for
#      correlation nor population noise

       write.morpho(  M = C, filename = "seqfile.aln" )

# A.2) Providing the morphological alignment (M = C), the population
#      noise (c = 0.25), and the name for the output file. Note that
#      c = 0.25 means that the population noise for all the characters
#      is c = 0.25, i.e. it will be considered as if
#      length( c ) = p characters, being all of them 0.25.

       write.morpho(  M = C, c = 0.25,
                      filename = "seqfile.aln" )

# A.3) Providing the morphological alignment (M = C), the population
#      noise (c = 0.25), the (estimate of the) correlation matrix (R),
#      the method to decompose R ("chol" in this example),
#      and the name for the output file. Note that the R matrix needs
#      to be invertible, otherwise the data will not be able to be
#      transformed accounting for correlation.

       write.morpho(  M = C, c = 0.25, R = R.sh,
                      method = "chol", filename = "seqfile.aln" )

# A.4) Providing the morphological alignment (M = C), a vector with
#      the population noise for each character (c = var.foxes),
#      the (shrinkage estimate of the) correlation matrix (R = R.sh),
#      the method to decompose R ("chol" in this example),
#      and the name for the output file. Note that the matrix passed
#      to argument R needs to be invertible, otherwise the data will
#      not be able to be transformed accounting for correlation.

       write.morpho(  M = C, c = var.foxes, R = R.sh,
                      method = "chol", filename = "seqfile.aln" )

# A.5) Providing the morphological alignment (C), a vector with
#      the population noise for each character (c = var.foxes),
#      the (shrinkage estimate of the) correlation matrix (R = R.sh),
#      the A matrix to transform the data, and the name for the
#      output file. Note that as the A matrix is provided, the matrix
#      passed to R will not be decomposed, hence the argument "method"
#      is not needed.

       write.morpho(  M = C, c = var.foxes, R = R.sh,
                      A = A, filename = "seqfile.aln" )

# B) Scenario A.5 but providing a list with the
#    names of the species

     names <- list( sp1  = "Ael_sp.", sp2  = "Can_dir", sp3  = "Epi_hay", sp4  = "Hes_sp.",
                    sp5  = "Par_jos", sp6  = "Tom_sp.", sp7  = "Enh_pah", sp8  = "Cuo_alp",
                    sp9  = "Spe_ven", sp10 = "Can_lup", sp11 = "Cer_tho", sp12 = "Oto_meg",
                    sp13 = "Urs_ame", sp14 = "Ail_ful", sp15 = "Nan_bin", sp16 = "Par_her",
                    sp17 = "Hia_won", sp18 = "Smi_fat", sp19 = "Vul_vul"
                  )

     write.morpho( M = C, c = var.foxes, R = R.sh,
                   A = A, filename = "seqfile.aln", names = names )

# C) Scenario A.5 but providing a vector of type character with the names of
#    the specimens and a list with their corresponding ages. Please
#    keep the same order in both lists, so the first specimen in the
#    list name corresponds to the first age in the age list, and so on.

     names <- c( "Ael_sp.", "Can_dir", "Epi_hay", "Hes_sp.",
                 "Par_jos", "Tom_sp.", "Enh_pah", "Cuo_alp",
                 "Spe_ven", "Can_lup", "Cer_tho", "Oto_meg",
                 "Urs_ame", "Ail_ful", "Nan_bin", "Par_her",
                 "Hia_won", "Smi_fat", "Vul_vul"
                  )

     ages <- list( sp1  = 13.135, sp2  =  0.0285, sp3  = 11.95,  sp4  = 35.55,
                   sp5  = 25.615, sp6  = 14.785,  sp7  = 28.55,  sp8  =  0,
                   sp9  =  0,     sp10 =  0,      sp11 =  0,     sp12 =  0,
                   sp13 =  0,     sp14 =  0,      sp15 =  0,     sp16 =  0,
                   sp17 =  6.65,  sp18 =  0.0285, sp19 =  0
                  )

     write.morpho( M = C, c = var.foxes, R = R.sh,
                   A = A, filename = "seqfile.aln",
                   names = names, ages = ages )

}
\seealso{
\code{\link{matrix2array}}, \code{\link{array2matrix}}, \code{\link{sim.morpho}}
}
\author{
Sandra Alvarez-Carretero and Mario dos Reis
}
