% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.R
\name{write.morpho}
\alias{write.morpho}
\title{Generate a phylip file for MCMCTree}
\usage{
write.morpho(M, filename, c = 0, R = diag(1, dim(M)[2]), method, A, names,
  ages)
}
\arguments{
\item{M}{Matrix, 's' rows (specimens) and 'n' morphological
continuous characters(see details).}

\item{filename}{Character, name for the output file.}

\item{c}{Numeric, vector of population variances (see details).
If not provided, c = 0 (no population variance).}

\item{R}{Matrix, correlation matrix. Requires \code{method}
(see details). If not provided, R = I (no trait correlation).}

\item{method}{(Optional) character, either \code{"eigen"} or
\code{"col"}, method used to decompose the inverse of the shrunk
correlation matrix. Requires \code{R} (see details).}

\item{A}{(Optional) matrix, decomposed matrix. Requires \code{R}
but not \code{method} (see details).}

\item{names}{(Optional) list, species name included in the morphological
alignment (see examples B and C).}

\item{ages}{(Optional) list, ages of the species included in the
morpholical alignment (see example C).}
}
\description{
Generate an alignment file in phylip format for MCMCTree.
The option "seqfile" in the control file used by MCMCTree
should read the path to the file output by this function.
}
\details{
The matrix \code{X} has 's' rows, one for each specimen, and 'n' columns regarding
the characters. If the data set contains landmarks, they can be given in 2D or 3D.
For instance, if the landmarks are 3D, the first 3 columns will be the
coordinates x, y, and z for the first landmark, the next 3
columns for the second landmark, and so on.
\tabular{cccccccc}{
 specimens \tab lmk1.x \tab lmk1.y  \tab lmk1.z  \tab lmk2.x \tab lmk2.y \tab lmk2.z  \tab ... \cr
 Sp_1      \tab 0.143  \tab -0.028  \tab -0.044  \tab 0.129  \tab 0.028  \tab -0.043  \tab ... \cr
 Sp_2      \tab 0.128  \tab -0.024  \tab -0.028  \tab 0.124  \tab 0.027  \tab -0.025  \tab ... \cr
 ...       \tab ...    \tab ...     \tab ...     \tab ...    \tab ...    \tab ...     \tab ...
}
See \code{canids19x29.matrix} in the \code{data} directory for a
more detailed example and \code{morpho/data-raw/canids19x29.matrix.R} for
details on how to generate this object from the raw data.
If the data set contains a set of 'n' morphological continuous characters,
e.g. from a simulated data set, the file should look like
\tabular{cccccc}{
 specimens \tab trait.1 \tab trait.2  \tab trait.2  \tab trait.2 \tab ... \cr
 Sp_1      \tab 0.143   \tab -0.028   \tab -0.044   \tab 0.129   \tab ... \cr
 Sp_2      \tab 0.128   \tab -0.024   \tab -0.028   \tab 0.124   \tab ... \cr
 ...       \tab ...     \tab ...      \tab ...      \tab ...     \tab ...
}

Note that if a list with the specimens names is not passed to the parameter \code{names},
the name for each species will be "Species_1", "Species_2", and so on.

The object \code{c} can be of length 1, if all characters have the same variance, or
a vector of length 'n' with the variance of each of the characters.

The object \code{R} has to be a symmetric and positive definite object of class matrix
(\code{class( R ) = "matrix"}).
The logarithm of the determinant of the correlation matrix is going to be printed
in the output file to later be used by MCMCTree during the likelihood calculation.

If a correlation matrix \code{R} is provided, \code{write.morpho()} can use either
the \code{method = "chol"} or \code{method = "eigen"} to get a matrix \code{A}
such that \eqn{\mathrm{R^{-1}}=\mathrm{A^{T}}\mathrm{A}}{R^{-1} = t(A) * A}.
This matrix \eqn{\mathrm{A^{T}}}{t(A)} is later used to transform
the morphological data while to account for the correlation in this data set,
so that the transformed characters in \code{Z},
\eqn{\mathrm{Z}=\mathrm{M}\mathrm{A^{T}}}{Z = M * t(A)}, are independent.
Alternatively, this matrix \code{A} can also be provided by the user. If so,
this matrix will be used to transform the data and no decomposition will be
performed, saving computational time for big matrices.
}
\examples{
# A.1) Providing the morphological alignment (M) and
#      the name for the output file. This does not account for
#      correlation nor population variance.

       write.morpho(  M = canids19x29.matrix, filename = "seqfile.aln" )

# A.2) Providing the morphological alignment (M), the population
#      variance (c), and the name for the output file. Note that
#      c = 0.25 means that the population variance for all the traits
#      is c = 0.25, i.e. it will be considered as if
#      length( c ) = p characters, being all of them 0.25.

       write.morpho(  M = canids19x29.matrix, c = 0.25,
                      filename = "seqfile.aln" )

# A.3) Providing the morphological alignment (M), the population
#      variance (c), the estimate of the correlation matrix (R),
#      the method to decompose R ("chol" in this example),
#      and the name for the output file. Note that the R matrix needs
#      to be invertible, otherwise the data will not be able to be
#      transformed accountign for correlation.

       write.morpho(  M = canids19x29.matrix, c = 0.25, R = R.shrunk,
                      method = "chol", filename = "seqfile.aln" )

# A.4) Providing the morphological alignment (M), the population
#      variance (c), the estimate of the correlation matrix (R),
#      the A matrix to transform the data, and the name for the
#      output file. Note that as the A matrix is provided, the R matrix
#      will not be decomposed, hence the argument "method" is no needed.

       write.morpho(  M = canids19x29.matrix, c = 0.25, R = R.shrunk,
                      A = A, filename = "seqfile.aln" )

# B) Scenario A.3 but providing a list with the
#    names of the species

     names <- list( sp1  = "Ael_sp.", sp2  = "Can_dir", sp3  = "Epi_hay", sp4  = "Hes_sp.",
                    sp5  = "Mes_cor", sp6  = "Tom_sp.", sp7  = "Enh_pah", sp8  = "Cuo_alp",
                    sp9  = "Spe_ven", sp10 = "Can_lup", sp11 = "Cer_tho", sp12 = "Oto_meg",
                    sp13 = "Vul_vul", sp14 = "Urs_ame", sp15 = "Ail_ful", sp16 = "Nan_bio",
                    sp17 = "Par_her", sp18 = "Tha_won", sp19 = "Smi_fat"
                  )

     write.morpho( M = canids19x29.matrix, c = 0.25, R = R.shrunk,
                   A = A, filename = "seqfile.aln", names = names )

# C) Scenario A.3 but providing a list with the names of
#    the specimens and a list with their corresponding ages. Please
#    keep the same order in both lists, so the first specimen in the
#    list name corresponds to the first age in the age list, and so on.

     names <- list( sp1  = "Ael_sp.", sp2  = "Can_dir", sp3  = "Epi_hay", sp4  = "Hes_sp.",
                    sp5  = "Mes_cor", sp6  = "Tom_sp.", sp7  = "Enh_pah", sp8  = "Cuo_alp",
                    sp9  = "Spe_ven", sp10 = "Can_lup", sp11 = "Cer_tho", sp12 = "Oto_meg",
                    sp13 = "Vul_vul", sp14 = "Urs_ame", sp15 = "Ail_ful", sp16 = "Nan_bio",
                    sp17 = "Par_her", sp18 = "Tha_won", sp19 = "Smi_fat"
                  )

     ages <- list( sp1  = 15.97, sp2  =  1.80, sp3  = 13.6,  sp4  = 39.74,
                   sp5  = 30.80, sp6  = 15.97, sp7  = 30.80, sp8  =  0,
                   sp9  =  0,    sp10 =  0,    sp11 =  0,    sp12 =  0,
                   sp13 =  0,    sp14 =  0,    sp15 =  0,    sp16 =  0,
                   sp17 =  0,    sp18 =  0,    sp19 =  0
                  )

     write.morpho( M = canids19x29.matrix, c = 0.25, R = R.shrunk,
                   A = A, filename = "seqfile.aln",
                   names = names, ages = ages )

}
\seealso{
\code{\link{matrix2array}}, \code{\link{array2matrix}}, \code{\link{sim.morpho}}
}
\author{
Sandra Alvarez-Carretero and Mario dos Reis
}
